Section1/pipeline/requirements.txt
pyspark==3.5.0
pyyaml==6.0.1
pytest==7.4.3
pytest-spark==0.6.0
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2

